merge

+++ EvergreenView/Controllers/MessageController.cs
@@ -1,4 +1,5 @@
 ﻿using EvergreenAPI.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
@@ -24,8 +25,8 @@ namespace EvergreenView.Controllers.Admin
             client = new HttpClient();
             var contentType = new MediaTypeWithQualityHeaderValue("application/json");
             client.DefaultRequestHeaders.Accept.Add(contentType);
-            MessageApiUrl = "https://localhost:44395/api/Message";
-            AccountApiUrl = "https://localhost:44395/api/Account";
+            MessageApiUrl = "https://localhost:44334/api/Message";
+            AccountApiUrl = "https://localhost:44334/api/Account";
             _httpContextAccessor = httpContextAccessor;
         }
 
@@ -40,12 +41,9 @@ namespace EvergreenView.Controllers.Admin
         }
 
 
-        public async Task<IActionResult> Index(string searchString)
+        public async Task<IActionResult> Index()
         {
-            if (session.GetString("User") == null || session.GetString("Role") != "Admin")
-            {
-                return RedirectToAction("Index", "Home");
-            }
+            
             string query = null;
            
             HttpResponseMessage response = await client.GetAsync(MessageApiUrl + query);
@@ -62,10 +60,7 @@ namespace EvergreenView.Controllers.Admin
         
         public async Task<ActionResult> Details(int id)
         {
-            if (session.GetString("User") == null || session.GetString("Role") != "Admin")
-            {
-                return RedirectToAction("Index", "Home");
-            }
+           
             var mess = await GetMessageByIdAsync(id);
             if (mess == null)
                 return NotFound();
@@ -74,11 +69,33 @@ namespace EvergreenView.Controllers.Admin
 
 
 
-        public async Task<ActionResult> Create()
+
+        [Authorize (Roles = "User")]
+        public async Task<ActionResult> UserCreate()
         {
-            if (session.GetString("User") == null || session.GetString("Role") != "Admin")
+            
+            HttpResponseMessage responeAccount = await client.GetAsync(AccountApiUrl);
+            string strData = await responeAccount.Content.ReadAsStringAsync();
+            var options = new JsonSerializerOptions
             {
-                return RedirectToAction("Index", "Home");
+                PropertyNameCaseInsensitive = true
+            };
+            return View();
+        }
+
+
+
+        [Authorize(Roles = "User")]
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> UserCreate(Message m)
+        {
+            string data = JsonSerializer.Serialize(m);
+            StringContent content = new StringContent(data, Encoding.UTF8, "application/json");
+            HttpResponseMessage response = client.PostAsync(MessageApiUrl, content).Result;
+            if (response.IsSuccessStatusCode)
+            {
+                return RedirectToAction("Index");
             }
             HttpResponseMessage responeAccount = await client.GetAsync(AccountApiUrl);
             string strData = await responeAccount.Content.ReadAsStringAsync();
@@ -87,14 +104,35 @@ namespace EvergreenView.Controllers.Admin
                 PropertyNameCaseInsensitive = true
             };
 
+            
+            return View();
+
+
+        }
+
+
+        [Authorize (Roles = "Admin")]
+        public async Task<ActionResult> AdminCreate()
+        {
+           
+            HttpResponseMessage responeAccount = await client.GetAsync(AccountApiUrl);
+            string strData = await responeAccount.Content.ReadAsStringAsync();
+            var options = new JsonSerializerOptions
+            {
+                PropertyNameCaseInsensitive = true
+            };
+
             List<Account> listAccount = JsonSerializer.Deserialize<List<Account>>(strData, options);
-            ViewData["Accounts"] = new SelectList(listAccount, "AccountId", "Username");
+            ViewData["Accounts"] = new SelectList(listAccount, "AccountId", "Fullname");
             return View();
         }
 
+
+
+        [Authorize (Roles = "Admin")]
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create(Message m)
+        public async Task<IActionResult> AdminCreate(Message m)
         {
             string data = JsonSerializer.Serialize(m);
             StringContent content = new StringContent(data, Encoding.UTF8, "application/json");
@@ -111,20 +149,18 @@ namespace EvergreenView.Controllers.Admin
             };
 
             List<Account> listAccount = JsonSerializer.Deserialize<List<Account>>(strData, options);
-            ViewData["Accounts"] = new SelectList(listAccount, "AccountId", "Username");
+            ViewData["Accounts"] = new SelectList(listAccount, "AccountId", "Fullname");
             return View();
 
 
         }
 
 
-        public async Task<ActionResult> Edit(int id)
-        {
-            if (session.GetString("User") == null || session.GetString("Role") != "Admin")
-            {
-                return RedirectToAction("Index", "Home");
-            }
 
+
+        [Authorize(Roles = "Admin")]
+        public async Task<ActionResult> AdminEdit(int id)
+        {
             HttpResponseMessage response = await client.GetAsync(MessageApiUrl + "/" + id);
             string strData = await response.Content.ReadAsStringAsync();
             var options = new JsonSerializerOptions
@@ -137,14 +173,17 @@ namespace EvergreenView.Controllers.Admin
             string str = await responeAccount.Content.ReadAsStringAsync();
 
             List<Account> listAccount = JsonSerializer.Deserialize<List<Account>>(str, options);
-            ViewData["Accounts"] = new SelectList(listAccount, "AccountId", "Username");
+            ViewData["Accounts"] = new SelectList(listAccount, "AccountId", "Fullname");
             return View(message);
         }
 
 
+
+
+        [Authorize(Roles ="Admin")]
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Edit(int id, Message product)
+        public async Task<ActionResult> AdminEdit(int id, Message product)
         {
             string data = JsonSerializer.Serialize(product);
             StringContent content = new StringContent(data, Encoding.UTF8, "application/json");
@@ -161,16 +200,84 @@ namespace EvergreenView.Controllers.Admin
             };
 
             List<Account> listAccounts = JsonSerializer.Deserialize<List<Account>>(strData, options);
-            ViewData["Accounts"] = new SelectList(listAccounts, "AccountId", "Username");
+            ViewData["Accounts"] = new SelectList(listAccounts, "AccountId", "Fullname");
+            return View();
+
+        }
+
+
+        [Authorize(Roles = "User")]
+        public async Task<ActionResult> UserEdit(int id)
+        {
+            var accountId = session.GetString("AccountId");
+           if (accountId == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+           
+           
+
+            HttpResponseMessage response = await client.GetAsync(MessageApiUrl + "/" + id);
+            string strData = await response.Content.ReadAsStringAsync();
+            var options = new JsonSerializerOptions
+            {
+                PropertyNameCaseInsensitive = true
+            };
+            Message message = JsonSerializer.Deserialize<Message>(strData, options);
+
+            if(accountId != message.AccountId.ToString())
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
+            HttpResponseMessage responeAccount = await client.GetAsync(AccountApiUrl);
+            string str = await responeAccount.Content.ReadAsStringAsync();
+
+            
+            return View(message);
+        }
+
+
+        [Authorize (Roles = "User")] 
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<ActionResult> UserEdit(int id, Message message)
+        {
+            var accountId = session.GetString("AccountId");
+            if (accountId == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+            if (accountId != message.AccountId.ToString())
+            {
+                return RedirectToAction("Index", "Home");
+            }
+            string data = JsonSerializer.Serialize(message);
+            StringContent content = new StringContent(data, Encoding.UTF8, "application/json");
+            HttpResponseMessage response = client.PutAsync(MessageApiUrl + "/" + id, content).Result;
+            if (response.IsSuccessStatusCode)
+            {
+                return RedirectToAction("Index");
+            }
+            HttpResponseMessage responeAccount = await client.GetAsync(AccountApiUrl);
+            string strData = await responeAccount.Content.ReadAsStringAsync();
+            var options = new JsonSerializerOptions
+            {
+                PropertyNameCaseInsensitive = true
+            };
+
+            
             return View();
 
         }
 
 
 
-        public async Task<ActionResult> Delete(int id)
+        [Authorize(Roles = "User")]
+        public async Task<ActionResult> UserDelete(int id)
         {
-            if (session.GetString("User") == null || session.GetString("Role") != "Admin")
+            var accountId = session.GetString("AccountId");
+            if (accountId == null)
             {
                 return RedirectToAction("Index", "Home");
             }
@@ -181,12 +288,26 @@ namespace EvergreenView.Controllers.Admin
             return View(mess);
         }
 
-
+        [Authorize(Roles = "User")]
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> UserDeleteConfirmed(int id)
         {
+            var accountId = session.GetString("AccountId");
+            if (accountId == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+           
+
+
+
             var message = await GetMessageByIdAsync(id);
+
+            if (accountId != message.AccountId.ToString())
+            {
+                return RedirectToAction("Index", "Home");
+            }
 
             HttpResponseMessage response = await client.DeleteAsync(MessageApiUrl + "/" + id);
             if (response.IsSuccessStatusCode)
@@ -222,7 +343,7 @@ namespace EvergreenView.Controllers.Admin
         public async Task SetViewData()
         {
             var listAccount = await GetAccountAsync();
-            ViewData["Accounts"] = new SelectList(listAccount, "AccountId", "Username");
+            ViewData["Accounts"] = new SelectList(listAccount, "AccountId", "Fullname");
         }
 
 
@@ -237,6 +358,36 @@ namespace EvergreenView.Controllers.Admin
 
             List<Account> listAccounts = JsonSerializer.Deserialize<List<Account>>(strData, options);
             return listAccounts;
+        }
+
+
+
+        [Authorize(Roles = "Admin")]
+        public async Task<ActionResult> AdminDelete(int id)
+        {
+ 
+            var mess = await GetMessageByIdAsync(id);
+            if (mess == null)
+                return NotFound();
+            await SetViewData();
+            return View(mess);
+        }
+
+        [Authorize(Roles = "Admin")]
+        [HttpPost, ActionName("Delete")]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> AdminDeleteConfirmed(int id)
+        {
+           
+            var message = await GetMessageByIdAsync(id);
+
+         
+            HttpResponseMessage response = await client.DeleteAsync(MessageApiUrl + "/" + id);
+            if (response.IsSuccessStatusCode)
+            {
+                return RedirectToAction("Index");
+            }
+            return View();
         }
     }
 }
